requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2

grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.18.0
six==1.16.0

paho-mqtt==1.5.1
psutil==5.8.0
